fix testing problems starting the server
